Fixed math pow

+++ libc/math/pow.c
@@ -1,10 +1,10 @@
 #include "math.h"
 
-int pow(int x, int y){
+double pow(double x, double y){
     if(y == 0){
         return 1;
     }
-    int x_p = x;
+    double x_p = x;
     while(y > 1){
         x_p *= x;
         --y;
